revert(agent-gui): remove session group pagination
