+++ apps/desktop/src/main/host/workspaceHostAccess.test.ts
@@ -290,9 +290,6 @@ function createTransportClient(
     async listWorkspaceAgentSessions() {
       throw new Error("not used");
     },
-    async listWorkspaceAgentSessionGroups() {
-      throw new Error("not used");
-    },
     async scanWorkspaceExternalAgentSessionImports() {
       throw new Error("not used");
     },
