+++ apps/desktop/src/main/host/workspaceLaunch.test.ts
@@ -407,10 +407,7 @@ function createTransportClient(
       return { terminals: [], workspaceId: workspaceID };
     },
     async listWorkspaceAgentSessions(workspaceID) {
-      return { hasMore: false, sessions: [], workspaceId: workspaceID };
-    },
-    async listWorkspaceAgentSessionGroups(workspaceID) {
-      return { groups: [], workspaceId: workspaceID };
+      return { sessions: [], workspaceId: workspaceID };
     },
     async scanWorkspaceExternalAgentSessionImports() {
       throw new Error("not used");
