+++ apps/desktop/src/renderer/src/features/workspace-agent/services/createDesktopAgentGUIWorkbenchHostInput.test.ts
@@ -2095,15 +2095,6 @@ function createWorkspaceAgentActivityService(
     async listAgentGeneratedFiles() {
       return { entries: [], workspaceId };
     },
-    async listSessionGroups() {
-      return { groups: [], workspaceId };
-    },
-    async listSessionsPage() {
-      return { hasMore: false, sessions: [], workspaceId };
-    },
-    async searchSessions() {
-      return { hasMore: false, sessions: [], workspaceId };
-    },
     async scanExternalSessionImports() {
       throw new Error("not implemented");
     },
