+++ apps/desktop/src/renderer/src/platform/tuttid/createDesktopTuttidClient.ts
@@ -504,12 +504,6 @@ export function createDesktopTuttidClient(
         request
       );
     },
-    async listWorkspaceAgentSessionGroups(workspaceID, request) {
-      return (await resolveClient()).listWorkspaceAgentSessionGroups(
-        workspaceID,
-        request
-      );
-    },
     async scanWorkspaceExternalAgentSessionImports(workspaceID, request) {
       return (await resolveClient()).scanWorkspaceExternalAgentSessionImports(
         workspaceID,
