+++ packages/agent/gui/agent-gui/agentGuiNode/AgentGUINode.spec.tsx
@@ -7237,13 +7237,6 @@ function createNoopAgentActivityRuntime(): AgentActivityRuntime {
     async listAgentGeneratedFiles(input) {
       return { workspaceId: input.workspaceId, entries: [] };
     },
-    async listSessionsPage(input) {
-      return {
-        workspaceId: input.workspaceId,
-        sessions: [],
-        hasMore: false
-      };
-    },
     async load(workspaceId) {
       return createAgentActivitySnapshotFromViewModel(workspaceId);
     },
