+++ packages/clients/tuttid-ts/src/tuttidClient.ts
@@ -45,7 +45,6 @@ import {
   listCliCapabilities,
   listWorkspaceAppMentionCandidates,
   listWorkspaceAgentGeneratedFiles,
-  listWorkspaceAgentSessionGroups,
   listUserProjects,
   listWorkspaceAgentSessionMessages,
   listWorkspaceIssues,
@@ -590,17 +589,6 @@ export function createTuttidClient(
         query: request
       });
       return unwrapData(response, "Workspace agent sessions request failed.");
-    },
-    async listWorkspaceAgentSessionGroups(workspaceID, request) {
-      const response = await listWorkspaceAgentSessionGroups({
-        client,
-        path: { workspaceID },
-        query: request
-      });
-      return unwrapData(
-        response,
-        "Workspace agent session groups request failed."
-      );
     },
     async listWorkspaceAgentGeneratedFiles(workspaceID, request) {
       const response = await listWorkspaceAgentGeneratedFiles({
