+++ services/tuttid/api/routes.go
@@ -354,14 +354,6 @@ func RegisterRoutes(mux *http.ServeMux, routes Routes) {
 		}
 	})
 
-	mux.HandleFunc("/v1/workspaces/{workspaceID}/agent-sessions/groups", func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != http.MethodGet {
-			tuttitypes.WriteMethodNotAllowed(w)
-			return
-		}
-		wrapper.ListWorkspaceAgentSessionGroups(w, r)
-	})
-
 	mux.HandleFunc("/v1/workspaces/{workspaceID}/agent-sessions/external-imports/scan", func(w http.ResponseWriter, r *http.Request) {
 		if r.Method != http.MethodPost {
 			tuttitypes.WriteMethodNotAllowed(w)
