+++ services/tuttid/service/agent/session_filter.go
@@ -16,9 +16,6 @@ func filterSessions(
 		if input.VisibleOnly && !session.Visible {
 			continue
 		}
-		if input.CWD != nil && strings.TrimSpace(session.Cwd) != strings.TrimSpace(*input.CWD) {
-			continue
-		}
 		if !matchesSessionSearch(session, input.SearchQuery) {
 			continue
 		}
